after much google, might have fixed arrival time/min left
wasnt sure if we were to assume the first train happened at some whatever time in the past? or the first train time was referring to the present day. made more sense to do present day, so i had to google fuu a lot and find similar projects that solved showing accurate next train and min left times assuming the first train time was for each day